Fixes #2.
Used hook to conditionally remove dabib.bib and/or makeidx.bat.
Also used conditionals in main.tex.

Condtionals are "smart" and will find the options to be true for anything
containing "y" or "t" -- case insensitive -- as in "yes" and "true" and
all their variations.

Alternatively, the bibtex conditional could ask for the name of the bib
file instead. We'll see.

Closing this for now.